+++ kakao-shop/src/apis/Login/index.ts
@@ -1,6 +1,6 @@
 import { client } from '@apis/client';
 
-import { SignInRequest } from '@hooks/ui/useLoginForm';
+import { SignInRequest } from '@hooks/page/Login/useLoginForm';
 
 export const signIn = async ({ email, password }: Pick<SignInRequest, 'email' | 'password'>) => {
   const res = await client.post('/login', { email, password });
