+++ kakao-shop/src/components/atom/Carousel/BannerImageList.tsx
