+++ kakao-shop/src/components/atom/Carousel/BannerImageListItem.tsx
