+++ kakao-shop/src/components/atom/Carousel/lib/bannerImages.ts
