+++ kakao-shop/src/components/page/SignUp/Footer/index.tsx
